% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectFromTemplate.R
\name{createProjectFromTemplate}
\alias{createProjectFromTemplate}
\title{Create a new R project from template}
\usage{
createProjectFromTemplate(name, directorypath = NULL)
}
\arguments{
\item{name}{The name of you R project (it will be created a directory using this name).}

\item{directorypath}{The directory path in which the project will be created.}
}
\value{
An object list with handy information about the project, like path and initial setup.
}
\description{
This function creates an initial R project setup focused in data science.
}
\examples{
createProjectFromTemplate("Predictive-Analytics", "c:/temp")
}
\author{
Leandro Daniel
}
\keyword{data-science,}
\keyword{project,}
\keyword{setup}
