% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestAndTrainSamples.R
\name{createTestAndTrainSamples}
\alias{createTestAndTrainSamples}
\title{Create test and train samples}
\usage{
createTestAndTrainSamples(dataset, y_var)
}
\arguments{
\item{dataset}{The dataset you want to split into test and train samples.}
}
\description{
This function creates train and test datasets given a database.
}
\examples{
createTestAndTrainSamples(my_dataset, my_y_var)

}
\keyword{dataset}
