% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loansdefaulters-data.R
\docType{data}
\name{loansdefaulters}
\alias{loansdefaulters}
\title{Loans defaulters dataset}
\format{An object of class \code{data.frame} with 682 rows and 118 columns.}
\source{
\href{https://github.com/ldaniel/Predictive-Analytics}{Predictive-Analytics source}
}
\usage{
data(loansdefaulters)
}
\description{
Dataset developed from a set of models using a real anonymized Czech bank 
transactions, account info, and loan records released for PKDD'99 Discovery Challenge.
}
\examples{
mydataset <- data(loansdefaulters)
}
\references{
Predictive-Analytics FGV's work assignment
(\href{https://ldaniel.github.io/Predictive-Analytics/}{Predictive-Analytics})
}
\keyword{datasets}
