% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestAndTrainSamples.R
\name{createTestAndTrainSamples}
\alias{createTestAndTrainSamples}
\title{Create test and train samples}
\usage{
createTestAndTrainSamples(dataset, yvar, seed = 12345,
  percentage = 0.7)
}
\arguments{
\item{dataset}{The dataset you want to split into test and train samples.}

\item{yvar}{The Y variable in your dataset.}

\item{seed}{The seed number used to generate the train and test samples.
The default value is 12345.}

\item{percentage}{The percentage of data that goes to training sample.
The default value is 0.7.}
}
\value{
An object list with the train dataset, test dataset and the proportions.
}
\description{
This function creates train and test datasets given a database and the Y variable.
}
\examples{
mydataset <- createTestAndTrainSamples(dataset = dataset, yvar = "Churn", seed = 12345, percentage = 0.7)
# or
mydataset <- createTestAndTrainSamples(dataset = dataset, yvar = "Churn")
# to use the final samples and to see the proportion
mydataset$data.train
mydataset$data.test
mydataset$event.proportion
}
\author{
Rodrigo Gonçalves and Leandro Daniel
}
\keyword{dataset}
